Move simulated example images to separate module
